+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for byteslice
 
+## 0.1.2.0 -- 2019-??-??
+
+* Add `Data.Bytes.Mutable` module.
+* Expose `unsafeTake` and `unsafeDrop` in `Data.Bytes`.
+
 ## 0.1.1.0 -- 2019-07-03
 
 * Add record labels for Bytes and MutableBytes